llm/pyproject.toml
[tool.poetry]
name = "llm"
version = "0.1.0"
description = ""
authors = ["jonirap <joni.rapoport@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index = "^0.6.25.post1"
langchain = "^0.0.199"
pygpt4all = "^1.1.0"
sentence-transformers = "^2.2.2"
accelerate = "^0.20.3"
gpt4all = "^0.3.4"
torch = {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-linux_x86_64.whl"}
hypsibius-slack = {path = "../hypsibius-slack", develop = true}
scikit-learn = "^1.2.2"
span-marker = "^1.2.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^1.3.0"
black = "^23.3.0"
notebook = "^6.5.4"
ipython = "^8.14.0"
tqdm = "^4.65.0"
ipywidgets = "^8.0.6"



[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
